// src/Models/api.ts
export interface ProcessHistoryDTO {
    /**
     * 
     * @type {string}
     * @memberof ProcessHistoryDTO
     */
    'action'?: string;
    /**
     * 
     * @type {string}
     * @memberof ProcessHistoryDTO
     */
    'actionStatus'?: ProcessHistoryDTOActionStatusEnum;
    /**
     * 
     * @type {string}
     * @memberof ProcessHistoryDTO
     */
    'comments'?: string;
    /**
     * 
     * @type {string}
     * @memberof ProcessHistoryDTO
     */
    'timestamp'?: string;
}

export const ProcessHistoryDTOActionStatusEnum = {
    Started: 'STARTED',
    Pending: 'PENDING',
    Completed: 'COMPLETED',
    Rejected: 'REJECTED',
    Cancelled: 'CANCELLED',
    Approved: 'APPROVED'
} as const;

export type ProcessHistoryDTOActionStatusEnum = typeof ProcessHistoryDTOActionStatusEnum[keyof typeof ProcessHistoryDTOActionStatusEnum];